% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_get_climdiv.R
\name{mco_get_climdiv}
\alias{mco_get_climdiv}
\title{Download the US Climate Division Dataset for Montana}
\usage{
mco_get_climdiv(elements = c("pcpn", "tmpc"))
}
\arguments{
\item{elements}{A character vector of elements to extract.\cr
pcpn = Precipitation\cr
tmpc = Average Temperature\cr
pdsi = Palmer Drought Severity Index\cr
phdi = Palmer Hydrological Drought Index\cr
zndx = Palmer Z Index\cr
pmdi = Modified Palmer Drought Severity Index\cr
hddc = Heating Degree Days\cr
cddc = Cooling Degree Days\cr
tmax = Maximum Temperature\cr
tmin = Minimum Temperature\cr
sp01 = 1-month Standardized Precipitation Index\cr
sp02 = 2-month Standardized Precipitation Index\cr
sp03 = 3-month Standardized Precipitation Index\cr
sp06 = 6-month Standardized Precipitation Index\cr
sp09 = 9-month Standardized Precipitation Index\cr
sp12 = 12-month Standardized Precipitation Index\cr
sp24 = 24-month Standardized Precipitation Index\cr}
}
\value{
A named data frame containing the climate division data.
\itemize{
\item \strong{Division code} — the NOAA Climate Divisional Database (nCLIMDIV) identifier
\item \strong{Element} — a four-sharacter string representing the climate element
\item \strong{Year} — the integer year
\item \strong{Month} — the integer month
\item \strong{Value} — the double precision element value
}
}
\description{
Download the US Climate Division Dataset for Montana
}
\examples{
\dontrun{
mco_get_climdiv("pcpn")
}
}
