% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_get_gridmet.R
\name{mco_get_gridmet}
\alias{mco_get_gridmet}
\title{Download select GridMET datasets.}
\usage{
mco_get_gridmet(x = mt_state \%>\% sf::st_buffer(10000),
  elements = c("precipitation_amount", "daily_minimum_temperature",
  "daily_maximum_temperature"), dates = "latest", out_dir = tempdir())
}
\arguments{
\item{x}{An object of class sf.}

\item{elements}{Gridmet variables to download.}

\item{dates}{An object of class Date or a character string formatted as
%Y-%m-%d (e.g., "2016-04-01") which specifies the date(s) to search.
To search for one specific date, this can be a Date object of length one. To
return data over a time interval, it can be a vector of length 2
(e.g., c("2018-01-01","2018-03-31")). Use "latest" (the default) to retreive the
most recent date available product.}

\item{out_dir}{A directory in which to download the raw MACA V2 datasets.
Defaults to the current working directory}
}
\value{
A raster brick of the desired MACA V2 datasets.
}
\description{
Download select GridMET datasets.
}
\examples{
\dontrun{
test <- mco_get_gridmet()
test <- mco_get_gridmet(dates = "2018-02-14")
test <- mco_get_gridmet(dates = "2018-03-31")
test <- mco_get_gridmet(dates = c("2018-01-01","2018-04-01"))
}
}
