% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_get_gridmet_normals.R
\name{mco_get_gridmet_normals}
\alias{mco_get_gridmet_normals}
\title{A wrapper around \link{mco_get_gridmet} to download GridMET normal dataset (1981--2010).
Here, the normals are defined as the daily mean values for the 1981--2010 period.}
\usage{
mco_get_gridmet_normals(x = mt_state \%>\% sf::st_buffer(10000),
  elements = c("precipitation_amount", "daily_minimum_temperature",
  "daily_maximum_temperature"), ...)
}
\arguments{
\item{x}{An object of class sf.}

\item{elements}{Gridmet variables to download.}

\item{...}{Additional arguments passed on to \link{mco_get_gridmet}.}
}
\value{
A raster brick of the desired MACA V2 dataset normals.
}
\description{
A wrapper around \link{mco_get_gridmet} to download GridMET normal dataset (1981--2010).
Here, the normals are defined as the daily mean values for the 1981--2010 period.
}
\examples{
\dontrun{
test <- mco_get_gridmet_normals()
}
}
