% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_get_macav2_monthly_datasets.R
\name{mco_get_macav2_monthly_datasets}
\alias{mco_get_macav2_monthly_datasets}
\title{Get a list of the available MACA V2 datasets.}
\usage{
mco_get_macav2_monthly_datasets()
}
\value{
A character vector of dataset names.
}
\description{
Get a list of the available MACA V2 datasets.
}
\examples{
mco_get_macav2_monthly_datasets()
}
