% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_get_smap.R
\name{mco_get_smap}
\alias{mco_get_smap}
\title{Download the NASA Soil Moisture Active Passive (SMAP) data for Montana}
\usage{
mco_get_smap(id = "SPL4SMGP", dates = "latest", raw_dir = NULL)
}
\arguments{
\item{id}{A character vector of SMAP dataset to extract.\cr
\describe{
\item{SPL3FTA}{Radar Northern Hemisphere Daily Freeze/Thaw State}
\item{SPL3SMA}{Radar Global Daily Soil Moisture}
\item{SPL3SMP}{Radiometer Global Soil Moisture}
\item{SPL3SMAP}{Radar/Radiometer Global Soil Moisture}
\item{SPL4SMAU}{Surface/Rootzone Soil Moisture Analysis Update}
\item{SPL4SMGP}{Surface/Rootzone Soil Moisture Geophysical Data}
\item{SPL4SMLM}{Surface/Rootzone Soil Moisture Land Model Constants}
\item{SPL4CMDL}{Carbon Net Ecosystem Exchange}
}}

\item{dates}{An object of class Date or a character string formatted as
%Y-%m-%d (e.g., "2016-04-01") which specifies the date(s) to search.
To search for one specific date, this can be a Date object of length one. To
search over a time interval, it can be a multi-element object of class Date
such as produced by \code{seq.Date}. Use "latest" (the default) to retreive the
most recent SMAP product.}

\item{raw_dir}{A directory in which to download the raw SMAP data. If missing,
data are downloaded to a temporary directory.}
}
\value{
A raster brick of the desired SMAP dataset.
}
\description{
This is a wrapper for \code{\link[smapr:find_smap]{smapr::find_smap()}} and \code{\link[smapr:download_smap]{smapr::download_smap()}}
that crops the resulting data to a ten kilometer buffer around the
state of Montana and allows the user to specify "latest" for the desired
\code{dates}, returning the most recent SMAP product.
}
\examples{
\dontrun{
mco_get_smap()
}
}
