% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_get_smap_appeears.R
\name{mco_get_smap_appeears}
\alias{mco_get_smap_appeears}
\title{Download the NASA Soil Moisture Active Passive (SMAP) data for Montana}
\usage{
mco_get_smap_appeears(id = "SPL4SMGP", group = "Geophysical_Data",
  name = "sm_rootzone", start_date = "latest", end_date = "latest",
  raw_dir = NULL)
}
\arguments{
\item{id}{A SMAP dataset to extract.\cr
\describe{
\item{SPL3FTA}{Radar Northern Hemisphere Daily Freeze/Thaw State}
\item{SPL3SMA}{Radar Global Daily Soil Moisture}
\item{SPL3SMP}{Radiometer Global Soil Moisture}
\item{SPL3SMAP}{Radar/Radiometer Global Soil Moisture}
\item{SPL4SMAU}{Surface/Rootzone Soil Moisture Analysis Update}
\item{SPL4SMGP}{Surface/Rootzone Soil Moisture Geophysical Data}
\item{SPL4SMLM}{Surface/Rootzone Soil Moisture Land Model Constants}
\item{SPL4CMDL}{Carbon Net Ecosystem Exchange}
}}

\item{group}{The SMAP dataset group name to extract.}

\item{name}{The SMAP dataset variable name to extract.}

\item{start_date}{An object of class Date or a character string formatted as
%Y-%m-%d (e.g., "2016-04-01") which specifies the start date of the download.}

\item{raw_dir}{A directory in which to download the raw SMAP data. If missing,
data are downloaded to a temporary directory.}

\item{start_date}{An object of class Date or a character string formatted as
%Y-%m-%d (e.g., "2016-04-01") which specifies the end date of the download.}
}
\value{
A raster brick of the desired SMAP dataset.
}
\description{
Download the NASA Soil Moisture Active Passive (SMAP) data for Montana
}
\examples{
\dontrun{
mco_get_smap_appeears()
}
}
