% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_get_snodas.R
\name{mco_get_snodas}
\alias{mco_get_snodas}
\title{Download and process the Montana SNOTEL site inventory from the
NRCS National Water and Climate Center}
\usage{
mco_get_snodas(
  x = mt_state \%>\% sf::st_buffer(10000),
  date = "latest",
  masked = TRUE,
  out_dir = tempdir(),
  overwrite = TRUE
)
}
\arguments{
\item{stations}{A character vector of station identifiers in trinomial format (e.g., '1040:CO:SNTL').}

\item{variables}{A character vector of variables to request.
Variables and their associated values are in variable/value pairs
in the format "VARIABLE::value" (for example, "WTEQ::normal_1981"
will give the 1981--2010 normal value for Snow Water Equivalent).
Possible variables include:
\itemize{
\item \strong{TAVG} --- Daily average air temperature (ºF)
\item \strong{TMAX} --- Daily maximum air temperature (ºF)
\item \strong{TMIN} --- Daily minimum air temperature (ºF)
\item \strong{TOBS} --- Instantaneously observed air temperature at start of day (ºF)
\item \strong{BATT} --- Battery voltage at start of day (volts)
\item \strong{BATX} --- Daily maximum battery voltage (volts)
\item \strong{PREC} --- Water year accumulated precipitation at start of day (inches)
\item \strong{PRCP} --- Daily total precipitation (inches)
\item \strong{PRCPSA} --- Daily total precipitation, snow-adjusted (inches)
\item \strong{SNWD} --- Total snow depth at start of day (inches)
\item \strong{WTEQ} --- Snow water equivalent at start of day (inches)
\item \strong{WTEQX} --- Daily maximum snow water equivalent (inches)
\item \strong{SNDN} --- Snow density at start of day (percent)
\item \strong{SNRR} --- Daily snow rain ratio (unitless)
}

Not all values are available for all variables. Possible values include:
\itemize{
\item \strong{value} --- The measured value
\item \strong{qcFlag} --- Quality control flag
\item \strong{qaFlag} --- Quality assurance flag
\item \strong{prevValue} --- Previous year's value
\item \strong{delta} --- Delta, change from previous year's value
\item \strong{collectionDate} --- Collection date
\item \strong{normal_1981} --- Normal (1981-2010)
\item \strong{pctOfNormal_1981} --- \% of normal (1981-2010)
\item \strong{average_1981} --- Average (1981-2010)
\item \strong{pctOfAverage_1981} --- \% of average (1981-2010)
\item \strong{median_1981} --- Median (1981-2010)
\item \strong{pctOfMedian_1981} --- \% of median (1981-2010)
\item \strong{normal_1971} --- Normal (1971-2000)
\item \strong{pctOfNormal_1971} --- \% of normal (1971-2000)
}}

\item{start_date}{A \link{Date} object as the start date. Defaults to period of record.}

\item{end_date}{A \link{Date} object as the end date. Defaults to period of record.}
}
\value{
A data frame containing the requested variables.
}
\description{
Download and process the Montana SNOTEL site inventory from the
NRCS National Water and Climate Center
}
\examples{
\dontrun{
mco_get_snotel_data()
}
}
