% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_get_snotel_data.R
\name{mco_get_snotel_data}
\alias{mco_get_snotel_data}
\title{Download and process the Montana SNOTEL site inventory from the
NRCS National Water and Climate Center}
\usage{
mco_get_snotel_data(stations = c("1040:CO:SNTL", "619:OR:SNTL",
  "1048:NM:SNTL"), variables = c("stationId", "WTEQ::value"),
  start_date = "POR_BEGIN", end_date = "POR_END")
}
\arguments{
\item{stations}{A character vector of station identifiers in trinomial format (e.g., '1040:CO:SNTL').}

\item{variables}{A character vector of variables to request.}

\item{start_date}{A \link{Date} object as the start date. Defaults to period of record.}

\item{end_date}{A \link{Date} object as the end date. Defaults to period of record.}
}
\value{
A data frame containing the requested variables.
}
\description{
Download and process the Montana SNOTEL site inventory from the
NRCS National Water and Climate Center
}
\examples{
\dontrun{
mco_get_snotel_data()
}
}
