% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_get_snotel_huc.R
\name{mco_get_snotel_huc}
\alias{mco_get_snotel_huc}
\title{Download and process the Montana SNOTEL Snow Water Equivalent dataset from the
NRCS National Water and Climate Center}
\usage{
mco_get_snotel_huc()
}
\value{
A data_frame with 2 fields:
\itemize{
\item \strong{Station} — the station identifier in trinomial format.
\item \strong{WBD code} — the WBD watershed code
}
}
\description{
Download and process the Montana SNOTEL Snow Water Equivalent dataset from the
NRCS National Water and Climate Center
}
\examples{
\dontrun{
mco_get_snotel_huc()
}
}
