% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_get_swe_normals.R
\name{mco_get_swe_normals}
\alias{mco_get_swe_normals}
\title{Download and process the Montana SNOTEL site inventory from the
NRCS National Water and Climate Center}
\usage{
mco_get_swe_normals(stations = NULL)
}
\arguments{
\item{stations}{A character vector of station identifiers in trinomial format (e.g., '1040:CO:SNTL').}
}
\value{
A simple feature collection with 4 fields:
\itemize{
\item \strong{Station Id} — the SNOTEL station identification number
\item \strong{Station Name} — the SNOTEL station location name
\item \strong{State Code} — the two letter state code
\item \strong{\verb{Network Code}} — the code for the station network
\item \strong{Start Date} — the date the stations started recording data
}
}
\description{
Download and process the Montana SNOTEL site inventory from the
NRCS National Water and Climate Center
}
\examples{
\dontrun{
mco_get_swe_normals()
}
}
