% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_leaflet.R
\name{mco_leaflet}
\alias{mco_leaflet}
\title{Basic leaflet plotting using \code{mcor} style}
\usage{
mco_leaflet(
  x,
  pal = "RdBu",
  legend_title = "",
  image_query_title = "",
  midpoint = 0,
  digits = 1,
  reverse = FALSE,
 
    attribution = "Climate data by <a href='http://www.climatologylab.org/gridmet.html' target='_blank'>gridMET</a> via <a href='https://earthengine.google.com/' target='_blank'>Google Earth Engine</a>"
)
}
\description{
Basic leaflet plotting using \code{mcor} style
}
