% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_leaflet.R
\name{mco_leaflet_base}
\alias{mco_leaflet_base}
\title{Basic leaflet plotting using \code{mcor} style}
\usage{
mco_leaflet_base(attribution = "")
}
\description{
Basic leaflet plotting using \code{mcor} style
}
