% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_mask.R
\name{mco_mask}
\alias{mco_mask}
\title{Mask a raster using fasterize}
\usage{
mco_mask(rast, mask)
}
\arguments{
\item{rast}{A Raster* object to be masked}

\item{mask}{A simple features polygon dataset to act as the mask.}
}
\value{
A masked Raster* object.
}
\description{
Mask a raster using fasterize
}
\examples{
\dontrun{
library(mcor)
library(magrittr)

raster::plot(mt_hillshade_500m)

# Mask to Missoula county
mco_mask(rast = mt_hillshade_500m,
         mask = dplyr::filter(mt_counties, County == "Missoula")) \%>\%
   raster::plot()
}
}
