% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_theme_map.R
\name{mco_theme_map}
\alias{mco_theme_map}
\title{Montana Climate Office theme for maps}
\usage{
mco_theme_map(base_size = 9, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
A \strong{ggplot2} theme that is good for displaying maps from \link[ggplot2:ggsf]{ggplot2::geom_sf}.
}
\examples{
\dontrun{
library(ggplot2)

ggplot(mt_counties_simple) +
  geom_sf(aes(fill = Division),
              color = "white") +
  mco_theme_map()
  }
}
