% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_climate_divisions.R
\docType{data}
\name{mt_climate_divisions}
\alias{mt_climate_divisions}
\title{Polygons of Montana Climate Divisions}
\format{A simple feature collection with 56 features and 3 fields:
\itemize{
\item \strong{Division} — the NOAA Climate Divisional Database (nCLIMDIV) climate division name
\item \strong{Division code} — the NOAA Climate Divisional Database (nCLIMDIV) identifier
\item \strong{Division FIPS code} — the Federal Information Processing Standard (FIPS) climate division code
}}
\source{
\url{https://data.nodc.noaa.gov/cgi-bin/iso?id=gov.noaa.ncdc:C00005}
}
\usage{
mt_climate_divisions
}
\description{
A dataset containing polygons of Montana climate divisions.
These data are derived from the \link{mt_counties} dataset;
county borders were dissolved using the \link[sf:dplyr]{sf::summarize.sf()} function.
This process ensures the extent of the state is the same as the \link{mt_counties} dataset.
These polygons are appropriate for large print maps.
The script that prepared the data is \url{data-raw/raw_data.R} in the source version of this
package.
}
\keyword{datasets}
