% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_counties_simple.R
\docType{data}
\name{mt_counties_simple}
\alias{mt_counties_simple}
\title{Simplified polygons of Montana Counties}
\format{A simple feature collection with 56 features and 8 fields:
\itemize{
\item \strong{STATE_FIPS} — the Federal Information Processing Standard (FIPS) state code
\item \strong{COUNTY_FIPS} — the Federal Information Processing Standard (FIPS) county code
\item \strong{COUNTY_ANSI} — the American National Standards Institute (ANSI) county code
\item \strong{COUNTY_GEOID} — the US Geographic Identifier (GEOID) county code
\item \strong{COUNTY_NAME} — the county name
\item \strong{CLIMATE_DIVISION_ID} — the NOAA Climate Divisional Database (nCLIMDIV) identifier
\item \strong{CLIMATE_DIVISION_FIPS} — the Federal Information Processing Standard (FIPS) climate division code
\item \strong{CLIMATE_DIVISION_NAME} — the NOAA Climate Divisional Database (nCLIMDIV) climate division name
}}
\source{
\url{https://www2.census.gov/geo/tiger/TIGER2017/COUNTY/tl_2017_us_county.zip}
}
\usage{
mt_counties_simple
}
\description{
A dataset containing polygons of Montana counties.
These data are \link{mt_counties}, simplified using \code{\link[rmapshaper:ms_simplify]{rmapshaper::ms_simplify()}}.
These polygons are appropriate for web maps and small print maps.
The script that prepared the data is \url{data-raw/raw_data.R} in the source version of this
package.
}
\keyword{datasets}
