% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_state_simple.R
\docType{data}
\name{mt_state_simple}
\alias{mt_state_simple}
\title{Simplified Polygon of Montana Boundary}
\format{
A geometry set with one feature.
}
\source{
\url{http://ftp.geoinfo.msl.mt.gov/Data/Spatial/MSDI/AdministrativeBoundaries/MontanaCounties.zip}
}
\usage{
mt_state_simple
}
\description{
A dataset containing the Montana boundary polygon.
These data are \link{mt_counties}, simplified using \code{\link[rmapshaper:ms_simplify]{rmapshaper::ms_simplify()}}
and then dissolved using the \link[sf:geos_combine]{sf::st_union} function.
These polygons are appropriate for web maps and small print maps.
The script that prepared the data is \url{data-raw/raw_data.R} in the source version of this
package.
}
\keyword{datasets}
