% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_tribal_land.R
\docType{data}
\name{mt_tribal_land}
\alias{mt_tribal_land}
\title{Polygons of Montana Tribal Land}
\format{
A simple feature collection with 7 features and 1 field:
\itemize{
\item \strong{Name} — the reservation name
}
}
\source{
\url{http://ftp.geoinfo.msl.mt.gov/Data/Spatial/MSDI/AdministrativeBoundaries/MontanaReservations.zip}
}
\usage{
mt_tribal_land
}
\description{
A dataset containing polygons of Montana Tribal Lands.
These data are available from \href{http://ftp.geoinfo.msl.mt.gov/Data/Spatial/MSDI/AdministrativeBoundaries/MontanaReservations.zip}{the Montana State Library},
and transformed using \code{\link[sf:st_transform]{sf::st_transform()}} to the \emph{NAD 1983 HARN StatePlane Montana FIPS 2500} projection
coordinate system (\href{https://epsg.io/102300}{EPSG: 102300}).
The script that prepared the data is \url{data-raw/raw_data.R} in the source version of this
package.
}
\keyword{datasets}
