% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_watersheds.R
\docType{data}
\name{mt_watersheds}
\alias{mt_watersheds}
\title{The Watershed Boundary Dataset Intersecting Montana}
\format{A simple feature collection with 56 features and 3 fields:
\itemize{
\item \strong{WBD code} — the WBD identifier
\item \strong{Watershed} — the WBD watershed name
\item \strong{Hydrologic Unit} — the hydrologic unit representing the level of the polygon
}}
\source{
\url{https://nhd.usgs.gov/wbd.html}
}
\usage{
mt_watersheds
}
\description{
A dataset containing polygons of Watershed Boundary Dataset (WBD) that intersect Montana.
These data are simplified using \code{\link[rmapshaper:ms_simplify]{rmapshaper::ms_simplify()}}.
These polygons are appropriate for web maps and small and large print maps.
The script that prepared the data is \url{data-raw/raw_data.R} in the source version of this
package.
}
\keyword{datasets}
