% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_get_macav2_monthly.R
\name{mco_get_macav2_monthly}
\alias{mco_get_macav2_monthly}
\title{Download select MACA V2 datasets.}
\usage{
mco_get_macav2_monthly(
  x = mt_state \%>\% sf::st_buffer(10000),
  elements,
  models,
  scenarios,
  out_dir = tempdir(),
  ...
)
}
\arguments{
\item{x}{An object of class sf.}

\item{elements}{A character vector of climate elements.}

\item{models}{A character vector of datasets}

\item{scenarios}{A character vector of representative climate pathways}

\item{out_dir}{A directory in which to download the raw MACA V2 datasets.
Defaults to the current working directory}

\item{...}{Other parameters passed on to \link[thredds:tds_ncss_download]{thredds::tds_ncss_download}.}
}
\value{
A raster brick of the desired MACA V2 datasets.
}
\description{
Download select MACA V2 datasets.
}
\examples{
\dontrun{
mco_get_macav2_monthly(x = mt_state)
}
}
