% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mco_get_snotel_data.R
\name{mco_get_snotel_data}
\alias{mco_get_snotel_data}
\title{Download and process the Montana SNOTEL site inventory from the
NRCS National Water and Climate Center}
\usage{
mco_get_snotel_data(stations = c("1040:CO:SNTL", "619:OR:SNTL",
  "1048:NM:SNTL"), variables = c("stationId", "WTEQ::value"),
  start_date = "POR_BEGIN", end_date = "POR_END")
}
\value{
A simple feature collection with 6 fields:
\itemize{
\item \strong{Station Id} — the SNOTEL station identification number
\item \strong{Station Name} — the SNOTEL station location name
\item \strong{State Code} — the two letter state code
\item \strong{\code{Network Code}} — the code for the station network
\item \strong{Start Date} — the date the station started recording data
\item \strong{End Date} — the date the station stopped recording data.
Will be "2100-01-01" if station is still recording.
}
}
\description{
Download and process the Montana SNOTEL site inventory from the
NRCS National Water and Climate Center
}
\examples{
\dontrun{
mco_get_snotel_inventory()
}
}
