% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_counties_simple.R
\docType{data}
\name{mt_counties_simple}
\alias{mt_counties_simple}
\title{Simplified polygons of Montana Counties}
\format{A simple feature collection with 56 features and 8 fields:
\itemize{
\item \strong{County} — the county name
\item \strong{State FIPS code} — the Federal Information Processing Standard (FIPS) state code
\item \strong{County FIPS code} — the Federal Information Processing Standard (FIPS) county code
\item \strong{Division} — the NOAA Climate Divisional Database (nCLIMDIV) climate division name
\item \strong{Division code} — the NOAA Climate Divisional Database (nCLIMDIV) identifier
\item \strong{Division FIPS code} — the Federal Information Processing Standard (FIPS) climate division code
}}
\source{
\url{http://ftp.geoinfo.msl.mt.gov/Data/Spatial/MSDI/AdministrativeBoundaries/MontanaCounties.zip}
}
\usage{
mt_counties_simple
}
\description{
A dataset containing polygons of Montana counties.
These data are \link{mt_counties}, simplified using \code{\link[rmapshaper:ms_simplify]{rmapshaper::ms_simplify()}}.
These polygons are appropriate for web maps and small print maps.
The script that prepared the data is \url{data-raw/raw_data.R} in the source version of this
package.
}
\keyword{datasets}
